LPS-57730 Move upgrades to module
